DOER-28 added task functionality completed

+++ backend/src/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, Column, Integer, String
+from sqlalchemy import Boolean, Column, Integer, String, BigInteger
 from .database import Base
 
 class Task(Base):
@@ -7,7 +7,7 @@ class Task(Base):
     title = Column(String, index=True)
     tags = Column(String, default=None)
     description = Column(String)
-    timestamp = Column(Integer)
+    timestamp = Column(BigInteger)
     deadline = Column(String, default=None)
     priority = Column(String, default=None)
     status = Column(String, default="To Do")
